Track both Upgrade and and Upgrade > Plan click event
