+++ client/my-sites/sidebar/item.jsx
@@ -6,16 +6,12 @@
  * These two cases might be to be split up?
  */
 
-import { isWooExpressPlan, PLAN_ECOMMERCE_TRIAL_MONTHLY } from '@automattic/calypso-products';
 import PropTypes from 'prop-types';
 import { memo } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import SidebarCustomIcon from 'calypso/layout/sidebar/custom-icon';
 import SidebarItem from 'calypso/layout/sidebar/item';
-import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { collapseAllMySitesSidebarSections } from 'calypso/state/my-sites/sidebar/actions';
-import { getSitePlanSlug } from 'calypso/state/sites/selectors';
-import { getSelectedSiteId } from 'calypso/state/ui/selectors';
 import MySitesSidebarUnifiedStatsSparkline from './sparkline';
 
 export const MySitesSidebarUnifiedItem = ( {
@@ -29,35 +25,19 @@ export const MySitesSidebarUnifiedItem = ( {
 	url,
 	shouldOpenExternalLinksInCurrentTab,
 	canNavigate,
+	trackClickEvent,
 } ) => {
 	const reduxDispatch = useDispatch();
-	const selectedSiteId = useSelector( getSelectedSiteId );
-	const sitePlanSlug = useSelector( ( state ) => getSitePlanSlug( state, selectedSiteId ) );
-
-	const trackItemClick = () => {
-		// For now, we only track clicks on the Plans menu item for WooExpress sites.
-		const isEcommerceTrial = sitePlanSlug === PLAN_ECOMMERCE_TRIAL_MONTHLY;
-		if ( ! isEcommerceTrial && ! isWooExpressPlan( sitePlanSlug ) ) {
-			return;
-		}
-
-		if ( url.startsWith( '/plans/' ) ) {
-			// Note that we also track this event in WooCommerce Screen via wc-calypso-bridge. If you change this event, please update it there as well. See: https://github.com/Automattic/wc-calypso-bridge/pull/1156.
-			reduxDispatch(
-				recordTracksEvent( 'calypso_sidebar_item_click', {
-					path: '/plans',
-				} )
-			);
-		}
-	};
 
 	const onNavigate = ( event ) => {
 		if ( ! canNavigate( url ) ) {
 			event?.preventDefault();
 			return;
 		}
+		if ( typeof trackClickEvent === 'function' ) {
+			trackClickEvent( url );
+		}
 
-		trackItemClick();
 		reduxDispatch( collapseAllMySitesSidebarSections() );
 		window.scrollTo( 0, 0 );
 	};
