+++ client/my-sites/sidebar/menu.jsx
@@ -5,14 +5,18 @@
  * This item can be expanded and collapsed by clicking.
  */
 
+import { isWooExpressPlan, PLAN_ECOMMERCE_TRIAL_MONTHLY } from '@automattic/calypso-products';
 import { isWithinBreakpoint } from '@automattic/viewport';
 import PropTypes from 'prop-types';
 import { useSelector, useDispatch } from 'react-redux';
 import SidebarCustomIcon from 'calypso/layout/sidebar/custom-icon';
 import ExpandableSidebarMenu from 'calypso/layout/sidebar/expandable';
 import { PromoteWidgetStatus, usePromoteWidget } from 'calypso/lib/promote-post';
+import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { toggleMySitesSidebarSection as toggleSection } from 'calypso/state/my-sites/sidebar/actions';
 import { isSidebarSectionOpen } from 'calypso/state/my-sites/sidebar/selectors';
+import { getSitePlanSlug } from 'calypso/state/sites/selectors';
+import { getSelectedSiteId } from 'calypso/state/ui/selectors';
 import MySitesSidebarUnifiedItem from './item';
 import { itemLinkMatches } from './utils';
 
@@ -33,6 +37,8 @@ export const MySitesSidebarUnifiedMenu = ( {
 	const reduxDispatch = useDispatch();
 	const sectionId = 'SIDEBAR_SECTION_' + slug;
 	const isExpanded = useSelector( ( state ) => isSidebarSectionOpen( state, sectionId ) );
+	const selectedSiteId = useSelector( getSelectedSiteId );
+	const sitePlanSlug = useSelector( ( state ) => getSitePlanSlug( state, selectedSiteId ) );
 	const selectedMenuItem =
 		Array.isArray( children ) &&
 		children.find( ( menuItem ) => menuItem?.url && itemLinkMatches( menuItem.url, path ) );
@@ -44,6 +50,32 @@ export const MySitesSidebarUnifiedMenu = ( {
 		( isDesktop && childIsSelected && ! sidebarCollapsed ); // For desktop breakpoints, a child should be selected and the sidebar being expanded.
 
 	const shouldShowAdvertisingOption = usePromoteWidget() === PromoteWidgetStatus.ENABLED;
+
+	const trackClickEvent = ( _link ) => {
+		// For now, we only track clicks on the Plans menu item for WooExpress sites.
+		const isEcommerceTrial = sitePlanSlug === PLAN_ECOMMERCE_TRIAL_MONTHLY;
+		if ( ! isEcommerceTrial && ! isWooExpressPlan( sitePlanSlug ) ) {
+			return;
+		}
+
+		if ( typeof _link !== 'string' || ! _link.startsWith( '/plans/' ) ) {
+			return;
+		}
+
+		// Check if we're navigating to /plans/:siteSlug or some deeper path below /plans/[something]/siteSlug
+		// The implementation can be changed to be simpler or different, but the check is needed.
+		const hasDeeperPath = _link.replace( '/plans/', '' ).includes( '/' );
+		if ( hasDeeperPath ) {
+			return;
+		}
+
+		// Note that we also track this event in WooCommerce Screen via wc-calypso-bridge. If you change this event, please update it there as well. See: https://github.com/Automattic/wc-calypso-bridge/pull/1156.
+		reduxDispatch(
+			recordTracksEvent( 'calypso_sidebar_item_click', {
+				path: '/plans',
+			} )
+		);
+	};
 
 	const onClick = ( event ) => {
 		// Block the navigation on mobile viewports and just toggle the section,
@@ -60,6 +92,7 @@ export const MySitesSidebarUnifiedMenu = ( {
 			return;
 		}
 
+		trackClickEvent( link );
 		window.scrollTo( 0, 0 );
 		reduxDispatch( toggleSection( sectionId ) );
 	};
@@ -88,6 +121,7 @@ export const MySitesSidebarUnifiedMenu = ( {
 							key={ item.title }
 							{ ...item }
 							selected={ isSelected }
+							trackClickEvent={ trackClickEvent }
 							isSubItem={ true }
 							shouldOpenExternalLinksInCurrentTab={ shouldOpenExternalLinksInCurrentTab }
 							canNavigate={ canNavigate }
